Make template archives deterministic and stable when there are no source changes (#120)
When you run `npm run build`, the template zips were changed even if
there were no changes to the template source. This was super annoying
and I had to deal with way too many merge conflicts because of this.

This fixes it by using the built-in `zip` command instead of `archiver`,
which generates the same zip from the same source files. There might be
some differences between the Linux and macOS versions, we'll see after
the CI runs for the first time, so this might still not be perfect, but
let's see.