feat: add id to template manifest (#126)
I propose to replace the current template `name` with `id`. In this PR,
I'm only adding the IDs because by removing the names, I would break CLI
for anyone who's using the `apify create --template=abc` flag. I would
remove the `name` after some deprecation period.

Why is this useful?

1. The names don't follow any meaningful structure so you always have to
open the manifest to find the template name. The IDs follow a pattern so
they're a bit less random.
2. The name ID implies that changing them could break something.
Yesterday I had a chat with @jbartadev and we figured that changing the
`name` does not break anything, but today I learned that I would break
the CLI flags for people.
3. Given the structure of the IDs, there's much less need to rename
anything in the future, which should prevent future breaking changes.

In the CLI, I will add a new flag called `templateId` so the change
won't be breaking there and I'll deprecate the `template`. This whole
thing actually started with me not understanding what should go into
`template` and I learned that `template` is actually `templateName`.

---------

Co-authored-by: František Nesveda <frantisek@apify.com>