fix: Add 60 minutes timeout to template tests (#292)
The broken template tests cost us $500 in the last month, because the
tests were running for 6 hours before they timed out. This adds a 1-hour
time limit to them, so that next time they time out earlier and it's not
so costly.

I checked and the tests typically run for 30 minutes, so 1 hour timeout
should be enough.