feat: add Camoufox-based Crawlee JS/TS templates (#336)
Following the https://github.com/apify/apify-sdk-js/pull/364 and
https://github.com/apify/crawlee/pull/2842 , this PR adds
Camoufox-enabled templates to Apify Actor templates. The implementation
is heavily based on the existing Playwright + Chrome templates.

The only issue (I'm aware of) currently is the immense size of those
images (as they contain Chrome and we add Camoufox binaries). Installing
Camoufox directly to a `node-debian` image results in missing system
dependencies. While it might be possible to install those manually in
the Dockerfile, it might make the Dockerfile too complex for a regular
user.


![image](https://github.com/user-attachments/assets/fb0050fd-fadc-4bbc-80f3-0681dcfa2b92)