feat: Add html comments to the readme (#532)
In order to detect the readme wasn't changed by the user, we added a
comment which should be always present. In the future we will be
scanning for such comment in order to determine originality of the
readme.

Also added test to ensure new templates will have readmes as well